Fix timing issue in computation of vuln scan completion
In the previous implementation, it was possible that the number of received scan results was equal to the number of sent scan commands, BEFORE all scan commands for a given scan key had been processed. Scans were thus marked as COMPLETE, despite that not being the case.

We now track the completion in a database table. A row in the table is created before emitting events to Kafka, and it's initialized with the expected number of scan results.

As updates to the tracking table are done in serial fashion, thanks to Kafka's ordering guarantees, it's safe to use optimistic locking.

It also simplifies the Kafka Streams topology.

Signed-off-by: nscuro <nscuro@protonmail.com>