Make better use of producer batching when dispatching events to Kafka
`KafkaProducer` is batching records behind the scenes before sending them to the broker. If we wait for broker acknowledgement for every single event we sent, we never really allow batches to fill up, causing unnecessary blocking in our code, and more network round trips between producer and broker.

https://kafka.apache.org/34/javadoc/org/apache/kafka/clients/producer/KafkaProducer.html#send(org.apache.kafka.clients.producer.ProducerRecord,org.apache.kafka.clients.producer.Callback)

By dispatching events asynchronously, batches are more likely to fill up before being submitted. This is especially beneficial when dispatching lots of events in a short period of time, for example after BOM uploads, or for vulnerability analysis of the entire portfolio.

If we enable snappy compression for the producer, it will be more effective because it applies to batches, not individual events.

Signed-off-by: nscuro <nscuro@protonmail.com>