Track the target entity of vulnerability scans
This allows us to, once a scan is deemed completed, decide if we need to kick off some other action(s) for only a component, or an entire project.

For policy evaluation and metrics updates specifically, it helps us in reducing load on API server and database by only executing them once for the entire project, instead of every single component within it.

Signed-off-by: nscuro <nscuro@protonmail.com>