Fix tests not being executed
A dependency transitively brings in junit-jupiter (junit 5), but all tests still use junit 4. Surefire prefers jupiter if it's present. We have to explicitly let surefire know that we want to execute tests using the legacy junit 4 engine.

This change was done in upstream, too. But I missed it when porting the Alpine update. https://github.com/DependencyTrack/dependency-track/pull/2530

Signed-off-by: nscuro <nscuro@protonmail.com>