Optimize querying of vulnerabilities and affected projects
Fetching all vulnerabilities via `/api/v1/vulnerability` would previously include too much data. Responses included vulnerability descriptions, credits, recommendations, affected components, and more. This caused queries to be extremely expensive, and API responses to be unnecessarily bloated, especially when a lot of components where affected by those vulnerabilities.

This has been addressed by introducing a new fetch group, which only includes "basic" fields, necessary to display an overview of vulnerabilities.

In order to enforce access control when portfolio ACL is enabled, DT would iterate over all affected components and their projects in-memory, evaluating whether the current principal is allowed to access them. In fact, this procedure is what caused component details to be populated in the API response. This was an unintended side effect.

This has been addressed by offloading most of the work to the database, which is achieved with a more sophisticated JDOQL query.

Signed-off-by: nscuro <nscuro@protonmail.com>