Switch to G1 garbage collector
With ParallelGC, we regularly see pauses in the >3 second range, which is simply too high for an application that's supposed to be responsive.

G1 is supposed to provide better latencies (through shorter GC pauses) than the parallel GC. Is is the default in Java >= 9 anyway.

It is also capable of de-duplicating strings in long-lived objects on the heap, which may lead to less heap usage overall.

Signed-off-by: nscuro <nscuro@protonmail.com>