Synchronously validate and persist uploaded BOMs to disk
... instead of validating them asynchronously, and passing them around entirely in-memory.

In situations where lots of BOMs are uploaded in close-ish succession, a large number of BOMs will pile up in memory as the API server struggles to keep up with processing them. This can cause memory limits of the JVM to be exceeded.

Additionally, in order to provide faster feedback to users uploading BOMs, validation is now performed before enqueueing the processing event.

Signed-off-by: nscuro <nscuro@protonmail.com>