Reduce volume of repo meta analysis commands sent
As repo meta analysis is entirely based on PURLs right now, and results are not associated with individual `Component`s, but `RepositoryMetaComponent`, there is no point in submitting an analysis command for every single `Component`.

Instead, we can de-duplicate based on `purlCoordinates` when querying the database. In portfolios with many shared components, this may result in a significant reduction in system load.

Signed-off-by: nscuro <nscuro@protonmail.com>