Skip unnecessary repartition of vuln scan results
The intention behind the repartition was to avoid race conditions when results for the same component would be processed in parallel. However, this case should be rather rare, and we already perform retries should a transaction fail due to a unique constraint violation.

It had the ugly side effect of doubling the partition count necessary to process vuln scan results, without adding much benefit.

Additionally, the re-keying of results to scan tokens also causes a repartition. This repartition included the scanner results, despite them not being needed by any of the following processing steps. To save some overhead, we now strip the scanner results before performing the repartition.

Signed-off-by: nscuro <nscuro@protonmail.com>