Avoid redundant copies of uploaded BOM content in-memory
The `BomParserFactory#looksLikeCycloneDX` accepts `byte[]`, and looks for certain CycloneDX markers in the provided value. It does so by converting the byte array to `String`, which internally will copy the entire array's content.

This is super inefficient for large BOMs, and not necessary at all. We already validate the file anyway.

Signed-off-by: nscuro <nscuro@protonmail.com>
Signed-off-by: mehab <meha.bhargava@citi.com>