Fix breaking change in `NEW_VULNERABILITY` notification JSON format
The breaking change was introduced for https://github.com/DependencyTrack/hyades/issues/467, and is causing issues for consumers of Webhook notifications that expect the `affectedProjects` field to be an array of projects. Whereas now it is an object holding references.

To fix the JSON representation, an array field named `affectedProjects` is added again. It will be populated with the same value that the `project` field holds.

Because only the field *name* is changed in the Protobuf schema, this is not a breaking change for the notification publisher.

Signed-off-by: nscuro <nscuro@protonmail.com>