Fix invalid Mattermost & Slack notification templates
Ported from https://github.com/DependencyTrack/dependency-track/pull/3095

Tests have not been ported, because the publishing of notifications now resides in the hyades notification-publisher, whereas the templates are still kept in the API server. It doesn't make sense to duplicate the template files to two repositories, risking that they become out-of-sync.

Signed-off-by: nscuro <nscuro@protonmail.com>