Fix FK violation during BOM processing
When uploading a BOM to an existing project, components that are no longer current are deleted. When to-be-deleted components had an `IntegrityAnalysis` associated with them, the deletion was failing due to a foreign key constraint violation.

This is now fixed by ensuring `IntegrityAnalysis` records are deleted prior to deleting the `Component` record.

Signed-off-by: nscuro <nscuro@protonmail.com>