Diff changes made by scanners to existing vulnerabilities
Instead of only logging *that* they were changed, also log which field was updated, and from which to which value.

Also reduce log level from `WARN` to `INFO`.

This is more or less a good-enough temporary solution, I realize we'll want similar functionality in various places throughout the code base.

Signed-off-by: nscuro <nscuro@protonmail.com>