Improve performance of `is_dependency_of` query
It is generally to be expected that the vast majority of components will *not* match the search criteria of `is_dependency_of`. Thus, it makes sense to optimize for the negative case.

The following additions were made:
* Perform a pre-flight check whether any component in the project matches the search criteria at all; The old query would traverse the entire graph even though no component could ever match
* Stop recursing once a match is found; The old query would keep note of when a component matched, but not abort the recursion, causing unnecessary queries to be performed

Queries where *no* component in the project matches the desired criteria complete in less than a millisecond.

Execution times of queries where at least one component matches the criteria now directly correlates with how "close" to the leaf node those components are.

Signed-off-by: nscuro <nscuro@protonmail.com>