Move vuln policy schema to `resources` (#482)
Non-Java files are not included in the final JAR, unless they're placed in `resources`. The schema file was not. This caused the file to be accessible in local testing, but not in the final Docker container.

Also named the schema and test files more explicitly as to not confuse it with other policy types.

Signed-off-by: nscuro <nscuro@protonmail.com>