Update component metrics in project metrics update procedure
Instead of fetching component UUIDs first, then calling the `UPDATE_COMPONENT_METRICS` procedure for every single UUID individually, and *then* updating the project metrics.

The previous approach resulted in:

* At least one `SELECT` query to fetch all N component UUIDs
* N calls to `UPDATE_COMPONENT_METRICS`
* N transactions, because each stored procedure call runs in its own transaction
* One call to `UPDATE_PROJECT_METRICS`

Which resulted in lots of wasted resources.

Refreshing metrics for all projects in a portfolio with 400 total projects completes in 5s instead of 20s on my machine. I expect the difference to be even more drastic in deployments where the database is running on a different machine, and network latency plays a bigger role.

Signed-off-by: nscuro <nscuro@protonmail.com>