Prevent invalid policy bundles from being processed
The original code did not assert the response code when downloading bundle files from HTTP servers (e.g. Nginx). Further, `ZipInputStream` does not fail when the `InputStream` passed to it produces bytes that are not a valid ZIP archive.

Response codes are now asserted, and bundle files are written to a temporary disk location. This allows us to use the `File`-based ZIP APIs, which do fail when the provided file is not a valid ZIP archive.

Signed-off-by: nscuro <nscuro@protonmail.com>