Handle direct dependency relationships in `is_exclusive_dependency_of`
Because the recursive query only operates on the `COMPONENT` table, it doesn't "see" occurrences of a component's UUID being referenced in `PROJECT.DIRECT_DEPENDENCIES`.

Added a separate query to check for this condition. It also acts as a short-circuiting mechanism that can avoid the expensive recursive query from being executed, as the existence of a direct dependency already false-ifies the "is exclusive dependency of another component" condition.

Signed-off-by: nscuro <nscuro@protonmail.com>