Recreate CPE parts index with computed `LOWER` columns
CPE matching must be case-insensitive as per CPE matching specification. We previously couldn't implement that, because defining indexes with computed columns was not possible in JDO.

As we're now using Liquibase, and switched the matching logic to use SQL instead of JDOQL / JPQL (https://github.com/DependencyTrack/hyades/pull/1213), we can finally resolve this properly.

Signed-off-by: nscuro <nscuro@protonmail.com>