Encode length constraints for vuln policy fields in JSON schema
The length of the `name`, `author`, and `description` fields are ultimately capped by the database schema. Until now, the JSON schema did not reflect those constraints, making it hard to detect invalid policies sooner.

This change extends the length of the `DESCRIPTION` column from `255` to `512`, and includes `minLength` / `maxLength` constraints for all string fields in the JSON schema.

Signed-off-by: nscuro <nscuro@protonmail.com>