Disable logging to file
For containerized workloads, logs are expected to be grabbed from STDOUT, writing them to disk unnecessarily adds storage requirements.

In the original config, three rolling file appenders, each with up to 10 files of up to 10MB each were configured. This led to up to 300MB of hidden storage requirement for long-running instances.

Signed-off-by: nscuro <nscuro@protonmail.com>