Fix `invalid stream header` when modifying analyses with associated vuln policy
When an existing analysis, which is assigned to a vulnerability policy, was updated, DataNucleus failed with `invalid stream header`. This happened because, when updating an analysis, DataNucleus fetches all associated objects, which includes `VulnerabilityPolicy`.

The `VulnerabilityPolicy` class does not map cleanly to the database schema: Some fields use PostgreSQL-native arrays, others use JSONB.

It also turned out that the model class isn't really used anywhere, since we modify and query `VulnerabilityPolicy` via JDBI. Hence the class was removed entirely, replacing it with a simple ID column of type `Long`.

Signed-off-by: nscuro <nscuro@protonmail.com>