Avoid redundant default object generation
Instead of re-running the default object generation, keep track of their version (currently, the build UUID of the application) in the `CONFIGPROPERTY` table. If the current application's build UUID matches that in the database, no action is performed.

This improved startup performance, since tasks like the synchronization of ~700 licenses is not cheap to do.

Signed-off-by: nscuro <nscuro@protonmail.com>