Fix metrics stored procs to only ever update the latest unchanged record
The previous version of the stored procedures had a bug where they would update the `LAST_OCCURRENCE` timestamp on existing records, as long as their values matched the calculated ones. It would do so despite the matching records not being the most recent.

As per the old Java-based implementation, only the most recent metrics entry is supposed to be re-used when all values match.

Signed-off-by: nscuro <nscuro@protonmail.com>