Fix mapping of `PURL` column to vulnerable components
JDBI's `BeanMapper` uses setters to map `ResultSet` columns to Java objects. Because the `purl` setter of the `Component` class exists twice with different argument types (once with `PackageURL` and once with `String`), JDBI picked the first, which failed because the `String` value of the `ResultSet` cannot be cast to the `PackageURL` type.

Switching to `FieldMapper` resolves the issue, since the `purl` field is of type `String`.

Signed-off-by: nscuro <nscuro@protonmail.com>