Fix `PortfolioMetricsUpdateTask` refreshing too many project metrics concurrently
In order to avoid a high impact on application and database, only up to `$CPU_CORE_COUNT` project metrics are supposed to be refreshed concurrently, when the `PortfolioMetricsUpdateTask` is running.

To achieve this, projects are supposed to be partitioned into at most `$CPU_CORE_COUNT` partitions. Unfortunately, the method used for partitioning divided the list of projects into N partitions of at most `$CPU_CORE_COUNT` elements.

This caused too many projects/partitions to be processed concurrently, claiming too many database connections of the connection pool.

I was unable to find an off-the-shelf implementation of the desired partitioning logic in any of the libraries we use, so built a custom one.

Signed-off-by: nscuro <nscuro@protonmail.com>