Fix PURL mismatch of INTEGRITY_META_COMPONENT records
When initializing records prior to event dispatch, the API server would use the PURL *coordinates* (i.e., excluding qualifiers etc.). However, analysis results arrive for the full PURL (which is correct).

When processing results, the API server failed to correlate the received result with the INTEGRITY_META_COMPONENT record it had created earlier.

As a result, integrity analysis did not work for component with PURLs that have qualifiers.

Signed-off-by: nscuro <nscuro@protonmail.com>