Disable DataNucleus bulk-fetching for list operations on `Project`
This only takes care of the worst offender for now, the underlying issue is still present and needs addressing.

Project listing is the worst offender because it currently causes the entire `PROJECTS_TAGS` table to fetched, even if the actual result sets of projects is only a tiny fraction of it.

Relates to https://github.com/DependencyTrack/hyades/issues/1754

Signed-off-by: nscuro <nscuro@protonmail.com>