Include Alpine JDO classes in `persistence.xml` of API Server
This fixes an issue where DataNucleus would not get to know about subclasses of `User`, until those subclasses are explicitly being used.

It turns out that classes defined in `alpine-model`'s `persistence.xml` are not implicitly included in `apiserver`'s `persistence.xml`.

For reference, this behavior is pointed out in the DN docs:

* https://www.datanucleus.org/products/accessplatform_6_0/jdo/mapping.html#inheritance
* https://www.datanucleus.org/products/accessplatform_6_0/jdo/persistence.html#autostart
* https://www.datanucleus.org/products/accessplatform_6_0/jdo/persistence.html#persistenceunit

Signed-off-by: nscuro <nscuro@protonmail.com>