Ingest datanucleus-postgresql plugin
Imports the https://github.com/nscuro/datanucleus-postgresql code. The plugin adds support for PostgreSQL JSONB methods in DataNucleus JDOQL, and UUIv7 value generation.

Since the plugin is only used by Dependency-Track, it makes sense to include it in the code base directly. It also reduces the bus factor, by moving it away from a repository only a single person (me) controls.

In a next step, plugin code in the `apiserver` module should be moved to `datanucleus-plugin` as well.

As part of this work, the version management of DataNucleus libraries and Maven plugins is centralized in the parent POM.

Signed-off-by: nscuro <nscuro@protonmail.com>