Separate workflow engine API into separate module
This is mostly to enforce a clearer distinction between public and internal API surface. We don't want DT to depend on engine internals, but only what we explicitly expose.

Adopt JPMS for workflow modules to make enforcement of this separation easier. The engine module does not export anything at all, except the `WorkflowEngineFactory` service implementation. Refer to https://dev.java/learn/modules/services/ for details on how this works.

The indirection of `WorkflowEngineFactory` is necessary because service providers must have a no-args constructor.

Signed-off-by: nscuro <nscuro@protonmail.com>