Fix test coverage diff not being displayed for PRs
It turns out that Codacy needs the `.git` directory to be present when it processes and uploads coverage reports. Without that, it won't match coverage info with files modified in the PR.

This updates the test workflow from only uploading coverage reports, to uploading the entire workspace, excluding a few well-known build artifacts.

To prevent the GitHub token from being leaked, the `checkout` action is configured to not store the token to begin with.

Signed-off-by: nscuro <nscuro@protonmail.com>