Fix `/api/v1/project/concise` performance regression
The latest metrics of each project were fetched in a lateral join. This causes the subquery to be executed for all rows, which got extremely expensive for a large number of projects (>100k).

Moving the metrics subquery to the `SELECT` statement will cause it to only execute for the actual result set (usually 10s or 100s or rows).

Unfortunately, this change means that metrics fields can no longer be sorted by.

Signed-off-by: nscuro <nscuro@protonmail.com>