Extend length of PURL columns to 1024
There are still occurrences in-the-wild where PURLs, or even only the PURL coordinates, can exceed the previous limit of 255 characters.

Since neither CycloneDX nor the PackageURL spec define a maximum acceptable length, truncating is not the way to go.

This adjusts the remaining PURL columns to accept values up to 1024 characters.

Signed-off-by: nscuro <nscuro@protonmail.com>