Pre-compute portfolio metrics for superusers or disabled ACL feature
Creates a materialized view that is refreshed whenever `PortfolioMetricsUpdateTask` runs. The refresh uses the `CONCURRENTLY` option, such that reads are not blocked while the refresh is still in progress.

This is a lot cheaper for cases where the portfolio spans thousands of projects, and the client has access to all of them. This can be the case if portfolio ACL is disabled, or the user has the `PORTFOLIO_ACCESS_CONTROL_BYPASS` permission.

Signed-off-by: nscuro <nscuro@protonmail.com>