Make portfolio metrics aggregation queries less expensive
The query previously considered the data of all previous days, for each day, when determining the latest data point.

This means that determining the latest data point for day 90 required sifting through 90 days worth of data.

The core requirement is to deal with gaps that may exist when projects don't have metrics for $today yet. It is sufficient to only consider $yesterday here, rather than the entire timespan.

Signed-off-by: nscuro <nscuro@protonmail.com>