Enable extensions to define types of their configs (#1397)
Previously, all configs were treated as strings, and extensions had to convert them on their own.

With this change, `ConfigDefinition`s can be typed, which centralizes the conversion logic and provides better type safety all around.

`ConfigType` being a sealed interface allows us to control which types are available. This is important for being able to render form inputs in the UI later. For example, `Boolean` would be a checkbox, `Instant` a time picker, and `Integer` a number input.

Signed-off-by: nscuro <nscuro@protonmail.com>