WIP: Remove WAR distribution
Instead of building a WAR and effectively deploying it to an embedded Jetty server, embed Jetty directly into the application JAR.

This makes it possible to simply launch the application via `main` method, and avoid the long WAR build times.

It is also easier to configure Jetty and customize it better for our specific requirements.

Signed-off-by: nscuro <nscuro@protonmail.com>