Bump container image to Java 25
Updates the JDK used to build the container image to Java 25, and enables the compact object headers feature (https://openjdk.org/jeps/519) for reduced CPU and memory overhead.

We will continue to build and test the application itself against Java 21 for the time being, and it will continue to be executable on Java 21 as well. This change purely affects the default image we distribute.

Signed-off-by: nscuro <nscuro@protonmail.com>