Migrate project cloning to SQL function
Cloning has previously been extremely expensive, because the entire project content had to be fetched and re-inserted without the ability to batch operations.

This made it necessary to offload cloning into an asynchronous task, which clients then had to poll for completion. Worse, clients did not get to know the UUID of the new project.

With this change, the cloning logic is migrated to a SQL function instead.

This way, database round-trips are minimized, batching is achieved, and performance is improved to a point that makes synchronous invocation feasible.

Signed-off-by: nscuro <nscuro@protonmail.com>