Implement transactional notification emission and dispatch
Introduces a mechanism to emit, route, and dispatch notifications without Kafka. It roughly follows the transactional outbox pattern.

Note that publishing (which requires interacting with external systems) is not part of this solution. It is going to be implemented using the upcoming workflow orchestration system (see ADR-002).

This implementation differs from the original ADR-003 in that it does not use PostgreSQL's logical replication. While that would have made the solution better performing, it would have required separate connection management, as well as elevated privileges (i.e., the `REPLICATION` role attribute).

Signed-off-by: nscuro <nscuro@protonmail.com>