Replace OR with IS DISTINCT FROM for suppression checks
It turns out that the PostgreSQL query planner yields suboptimal plans for `("SUPPRESSED" IS NULL OR NOT "SUPPRESSED")` conditions. `"SUPPRESSED" IS DISTINCT FROM TRUE` is semantically equivalent, avoids the `OR`, and causes the planner to generate better plans.

Eventually the `"SUPPRESSED"` column should be made non-null, but this change provides a quick relief without the need for an expensive migration.

Signed-off-by: nscuro <nscuro@protonmail.com>