Implement secret management
Adds a pluggable secret manager mechanism with basic create/update/get/list operations.

Defaults to a database-backed solution where secrets are encrypted and stored in the `SECRET` table.

Future implementations like Vault or AWS secret manager would allow users to deal with rotation of secrets without having to touch the Dependency-Track system.

Secret manager is not an extension point, because it is a core infrastructure component that will be used by the plugin system itself.

The main purpose of this feature is to allow other areas of the application to safely reference secrets. This will be achieved using templating.

Signed-off-by: nscuro <nscuro@protonmail.com>