Make workflow task queue partitions UNLOGGED
Workflow tasks are ephemeral and are trivially  recreated by the task scheduler. Tasks are scheduled based on whether new messages are available in a run's inbox, so the real source of truth that needs logging is the dex_workflow_inbox table.

Making the task table UNLOGGED saves a bunch of I/O overhead.

https://www.postgresql.org/docs/current/sql-createtable.html#SQL-CREATETABLE-UNLOGGED
Signed-off-by: nscuro <nscuro@protonmail.com>